requirements.txt
aniso8601==8.0.0
asgiref==3.4.1
astroid==2.7.3
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.7
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.1.0
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
idna==3.2
inflection==0.5.1
isort==5.9.3
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.21.2
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.0
pandas==1.3.2
pathspec==0.9.0
platformdirs==2.3.0
pylint==2.10.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2020.1
regex==2021.8.28
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.4.1
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
Werkzeug==1.0.1
wrapt==1.12.1
